+++ adapter/moloco/src/main/java/org/bidon/moloco/impl/MolocoInterstitialImpl.kt
@@ -4,9 +4,12 @@ import android.app.Activity
 import com.moloco.sdk.publisher.AdLoad
 import com.moloco.sdk.publisher.InterstitialAd
 import com.moloco.sdk.publisher.InterstitialAdShowListener
+import com.moloco.sdk.publisher.MediationInfo
 import com.moloco.sdk.publisher.Moloco
 import com.moloco.sdk.publisher.MolocoAd
 import com.moloco.sdk.publisher.MolocoAdError
+import org.bidon.moloco.EMPTY_MEDIATOR
+import org.bidon.moloco.EMPTY_WATERMARK
 import org.bidon.moloco.MolocoDemandId
 import org.bidon.moloco.ext.toBidonLoadError
 import org.bidon.moloco.ext.toBidonShowError
@@ -67,14 +70,7 @@ internal class MolocoInterstitialImpl :
 
             override fun onAdShowFailed(molocoAdError: MolocoAdError) {
                 logInfo(TAG, "onAdShowFailed: ${molocoAdError.description}")
-                emitEvent(
-                    AdEvent.ShowFailed(
-                        (BidonError.Unspecified(
-                            demandId,
-                            molocoAdError.toBidonShowError()
-                        ))
-                    )
-                )
+                emitEvent(AdEvent.ShowFailed(molocoAdError.toBidonShowError()))
             }
 
             override fun onAdHidden(molocoAd: MolocoAd) {
@@ -116,6 +112,8 @@ internal class MolocoInterstitialImpl :
             return
         }
         Moloco.createInterstitial(
+            mediationInfo = MediationInfo(EMPTY_MEDIATOR),
+            watermarkString = EMPTY_WATERMARK,
             adUnitId = adParams.adUnitId
         ) { interstitial: InterstitialAd?, adCreateError: MolocoAdError.AdCreateError? ->
             if (interstitial != null) {